% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_generation.R
\docType{data}
\name{first_generation_kinds}
\alias{first_generation_kinds}
\title{Available First Generation Kinds}
\format{
A character vector with values: "random", "grider", "segment10", "octagon".
}
\usage{
first_generation_kinds
}
\description{
A character vector of available pattern kinds for generating the first
generation of a cellular automaton grid.
}
\keyword{datasets}
