% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifegame.R
\name{lifegame}
\alias{lifegame}
\title{Lifegame}
\usage{
lifegame(input = first_generation(), n = 1, ...)
}
\arguments{
\item{input}{An matrix composed of \code{0L} and \code{1L}.}

\item{n}{Number of generations to calculate.}

\item{...}{Other parameters passed down to methods}
}
\value{
A lifegame class \code{data.frame} with 4 integer columns.
The \code{x} and \code{y} columns indicate their coordinates.
The \code{generation} column indicates n-th generation of lifegame.
The \code{life} column indicates if the cell is dead (\code{0}) or alive (\code{1L}).
}
\description{
Lifegame
}
