% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{blinker}
\alias{frog}
\alias{diehard}
\title{Toy datasets for lifegame}
\format{
An object of class \code{matrix} with 5 rows and 5 columns.

An object of class \code{matrix} with 6 rows and 6 columns.

An object of class \code{matrix} with 6 rows and 10 columns.
}
\usage{
blinker

frog

diehard
}
\description{
Toy datasets for lifegame
}
\keyword{datasets}
